pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud_tts"
version = "0.0.5"
authors = [
    { name="Ben Brenner", email="ben.brenner@gmail.com" },
    ]
description = "A tool for requesting text to speech from a variety of cloud APIs."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [

    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "azure-cognitiveservices-speech==1.29.0",
    "boto3==1.26.151",
    "google-cloud-texttospeech==2.14.1",
    "keyring==23.13.1",
]

[project.urls]
"Homepage" = "https://github.com/cobiadigital/cloudtts"
"Bug Tracker" = "https://github.com/cobiadigital/cloudtts/issues"


